+++ classes/Generator/Generator/Builder.php
@@ -18,7 +18,7 @@
 class Generator_Generator_Builder
 {
 	// Release version
-	const VERSION = '0.2';
+	const VERSION = '0.3';
 
 	// Generator commands
 	const CREATE = 'create';
@@ -139,11 +139,14 @@ class Generator_Generator_Builder
 	 * @param   array  $defaults   The list of default values
 	 * @return  Generator_Builder  This instance
 	 */
-	public function with_defaults(array $defaults)
+	public function with_defaults(array $defaults = NULL)
 	{
-		$this->_defaults = $defaults;
+		if ($defaults)
+		{
+			$this->_defaults = $defaults;
+			$this->_is_prepared = FALSE;
+		}
 
-		$this->_is_prepared = FALSE;
 		return $this;
 	}