+++ classes/Generator/Generator/Type/Config.php
@@ -29,28 +29,6 @@ class Generator_Generator_Type_Config extends Generator_Type
 	}
 
 	/**
-	 * Converts an array of value definition strings into a final array of 
-	 * value items.
-	 *
-	 * @param   array  $values  The list of value definitions
-	 * @return  array  The parsed list
-	 */
-	public function parse_values(array $values)
-	{
-		$ret = array();
-
-		foreach ($values as $value)
-		{
-			list($key, $val) = explode('|', $value);
-			$key = is_numeric($key) ? (trim($key) + 0) : trim($key);
-			$val = is_numeric($val) ? (trim($val) + 0) : trim($val);
-			Arr::set_path($ret, $key, $val);
-		}
-
-		return $ret;
-	}
-
-	/**
 	 * Finalizes parameters and renders the template.
 	 *
 	 * @return  string  The rendered output
@@ -59,10 +37,42 @@ class Generator_Generator_Type_Config extends Generator_Type
 	{
 		if ( ! empty($this->_params['values']))
 		{
-			$this->_params['values'] = $this->parse_values($this->_params['values']);
+			$this->_params['values'] = $this->_parse_values($this->_params['values']);
 		}
 
 		return parent::render();
+	}
+
+	/**
+	 * Converts an array of value definition strings into a final array of 
+	 * value items.
+	 *
+	 * @param   array  $values  The list of value definitions
+	 * @return  array  The parsed list
+	 */
+	protected function _parse_values(array $values)
+	{
+		$ret = array();
+
+		foreach ($values as $value)
+		{
+			list($key, $val) = explode('|', $value);
+			$key = is_integer($key) ? (trim($key) + 0) : trim($key);
+			$val = is_numeric($val) ? (trim($val) + 0) : trim($val);
+
+			if (is_numeric($key))
+			{
+				// Numeric keys should be set directly
+				$ret[$key] = $val;
+			}
+			else
+			{
+				// Otherwise treat key as an array path
+				Arr::set_path($ret, $key, $val);
+			}
+		}
+
+		return $ret;
 	}
 
 } // End Generator_Generator_Type_Config