+++ classes/Generator/Task/Generate/Task/Generator.php
@@ -34,22 +34,6 @@ class Generator_Task_Generate_Task_Generator extends Task_Generate
 	}
 
 	/**
-	 * Loads any view parameter defaults from config.
-	 *
-	 * @param  array  $options  The selected task options
-	 * @return array  The default values
-	 */
-	public function get_defaults(array $options = NULL)
-	{
-		$config = ! empty($options['config']) ? $options['config'] : 'generator';
-
-		if ($defaults = Kohana::$config->load($config.'.defaults.class'))
-			return $defaults;
-
-		return array();
-	}
-
-	/**
 	 * Creates a generator builder with the given configuration options.
 	 *
 	 * @param  array  $options  the selected task options
@@ -87,20 +71,8 @@ class Generator_Task_Generate_Task_Generator extends Task_Generate
 			->with_module($options['module'])
 			->with_pretend($options['pretend'])
 			->with_force($options['force'])
-			->with_defaults($this->get_defaults($options))
+			->with_defaults($this->get_config('defaults.class', $options['config']))
 			->prepare();
-	}
-
-	/**
-	 * Executes the task.
-	 *
-	 * @param  array  $params  The task parameters
-	 * @return void
-	 */
-	protected function _execute(array $params)
-	{
-		$builder = $this->get_builder($params);
-		$this->run($builder, $params);
 	}
 
 } // End Generator_Task_Generate_Task_Generator